analyzer.js에서 분석 로직과 출력 로직 분리
