+++ lib/analyzer.js
@@ -9,6 +9,8 @@ import Table from 'cli-table3';
 import {log, setTextColor} from './Util.js';
 import ThemeManager from './ThemeManager.js';
 import { ChartGenerator } from './ChartGenerator.js';
+import TableGenerator from './tableGenerator.js';
+import CsvGenerator from './csvGenerator.js';
 
 import { getBadge } from './Util.js';
 
@@ -377,130 +379,15 @@ class RepoAnalyzer {
         return newAllRepoScores;
     }
 
-    async generateTable(allRepoScores, output) {
-        const promises = Array.from(allRepoScores).map(async ([repoName, repoScores]) => {
-            let theme = this.themeManager.getCurrentTheme();
-            
-            // 테마 null 체크
-            if (!theme || !theme.table) {
-                console.error('테이블 테마를 불러올 수 없습니다. 기본 설정을 사용합니다.');
-                theme = {
-                    table: {
-                        head: ['yellow'],
-                        border: ['gray']
-                    }
-                };
-            }
-
-            // 한국어와 이모지의 실제 표시 너비 계산 함수
-            const calculateDisplayWidth = (str) => {
-                let width = 0;
-                for (const char of str) {
-                    // 한국어: 2바이트로 계산
-                    if (/[\u3131-\uD79D]/.test(char)) {
-                        width += 2;
-                    }
-                    // 이모지: 2바이트로 계산 (터미널마다 다를 수 있지만 기본적으로 2로 가정)
-                    else if (/[\uD800-\uDBFF][\uDC00-\uDFFF]/.test(char)) {
-                        width += 2;
-                    }
-                    // 그 외 ASCII 문자: 1바이트
-                    else {
-                        width += 1;
-                    }
-                }
-                return width;
-            };
-
-            // 첫 번째 열(참가자)의 최대 너비 계산
-            let maxNameWidth = 20; // 최소 너비
-            repoScores.forEach(([name, , , , , , total]) => {
-                const badge = getBadge(total);
-                // 뱃지에서 이모지만 추출
-                const badgeEmoji = badge.split(' ')[0];
-                const nameWithBadge = `${badgeEmoji} ${name}`;
-                const displayWidth = calculateDisplayWidth(nameWithBadge);
-                maxNameWidth = Math.max(maxNameWidth, displayWidth + 2); // 여유 공간 추가
-            });
-
-            // 동적 colWidths 설정
-            const colWidths = [
-                maxNameWidth, // 참가자 열 (동적)
-                16,           // feat/bug PR 점수
-                12,           // doc PR 점수
-                12,           // typo PR 점수
-                16,           // feat/bug 이슈 점수
-                12,           // doc 이슈 점수
-                10,           // 총점
-                11            // 참여율(%)
-            ];
-
-            const table = new Table({
-                head: ['참가자', 'feat/bug PR 점수', 'doc PR 점수', 'typo PR 점수', 'feat/bug 이슈 점수', 'doc 이슈 점수', '총점', '참여율(%)'],
-                colWidths: colWidths,
-                style: { 
-                    head: theme.table.head,
-                    border: theme.table.border
-                },
-                wordWrap: true // 긴 텍스트 자동 줄바꿈
-            });
-
-            // 리포지토리 전체 합(= 모든 기여자의 totalScore 합)
-            const totalScore = repoScores.reduce((sum, row) => sum + row[6], 0); // 변경: totalScore 인덱스 6으로 조정
-
-            // 텍스트 파일로 저장하기 위해 문자열 준비
-            repoScores.forEach(([name, p_fb_score, p_d_score, p_t_score, i_fb_score, i_d_score, total]) => { // 변경: p_t_score 추가
-                const rate = totalScore > 0 ? ((total / totalScore) * 100).toFixed(2) : '0.00';
-                //뱃지
-                const badge = getBadge(total);
-                // 뱃지에서 이모지만 추출
-                const badgeEmoji = badge.split(' ')[0];
-                const nameWithBadge = `${badgeEmoji} ${name}`;
-                // CLI에 표시될 테이블 행
-                table.push([
-                    nameWithBadge,
-                    p_fb_score,
-                    p_d_score,
-                    p_t_score, // 추가
-                    i_fb_score,
-                    i_d_score,
-                    total,
-                    `${rate}%`
-                ]);
-            });
-
-            const filePath = path.join(output, `${repoName}.txt`);
-            await fs.mkdir(output, { recursive: true });
-
-            const now = new Date();
-            const dateStr = now.toLocaleString('en-US', { timeZone: 'Asia/Seoul' });
-
-            const headerTextLines = [
-                'Contribution Score by Participant',
-                `Generated at ${dateStr}`,
-                '' // 줄바꿈 한 줄 추가
-            ];
-            const headerText = headerTextLines.join('\n');
-
-            // 테이블 문자열
-            const tableString = table.toString();
-            
-            // ANSI 색상 코드를 제거한 텍스트를 저장
-            // 정규식을 사용하여 ANSI 이스케이프 시퀀스 제거
-            const stripAnsi = (str) => {
-                // ANSI 이스케이프 시퀀스 제거 정규식
-                return str.replace(/\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])/g, '');
-            };
-
-            // 최종 출력 문자열: 헤더 + 테이블
-            const finalOutput = headerText + tableString;
-
-            // 파일로 저장
-            await fs.writeFile(filePath, stripAnsi(finalOutput), 'utf-8');
-            log(`점수 집계 텍스트 파일이 생성되었습니다: ${filePath}`, 'INFO');
-        });
-
-        await Promise.all(promises);
+    //테이블 생성
+    async generateTable(allRepoScores, outputDir = '.') {
+        try {
+            const tableGenerator = new TableGenerator(this.themeManager);
+            await tableGenerator.generateTable(allRepoScores, outputDir);
+        } catch (error) {
+            log(`테이블 생성 중 오류 발생: ${error.message}`, 'ERROR');
+            throw error;
+        }
     }
 
     /**
@@ -528,41 +415,15 @@ class RepoAnalyzer {
         }
     }
 
-    async generateCombinedCsv(repoName, repoScores, repoActivities, outputDir) {
-        let csvContent = CSV_HEADER;
-        let totalScore = 0;
-
-        // 참여자별로 점수와 횟수를 결합
-        for (const [participant, activities] of repoActivities.entries()) {
-            const p_fb_count = activities.pullRequests.bugAndFeat || 0;
-            const p_d_count = activities.pullRequests.doc || 0;
-            const p_t_count = activities.pullRequests.typo || 0;
-            const i_fb_count = activities.issues.bugAndFeat || 0;
-            const i_d_count = activities.issues.doc || 0;
-
-            // 해당 참여자의 점수 찾기
-            const scoreData = repoScores.find(([name]) => name === participant) || [participant, 0, 0, 0, 0, 0, 0];
-            const [, p_fb_score, p_d_score, p_t_score, i_fb_score, i_d_score, total] = scoreData;
-
-            csvContent += `${participant},${p_fb_count},${p_fb_score},${p_d_count},${p_d_score},${p_t_count},${p_t_score},${i_fb_count},${i_fb_score},${i_d_count},${i_d_score},${total}\n`;
-            totalScore += total;
-        }
-
-        const filePath = path.join(outputDir, `${repoName}_combined.csv`);
-        await fs.writeFile(filePath, csvContent);
-        log(`통합 CSV 파일이 생성되었습니다: ${filePath}`, 'INFO');
-    }
-
+    //CSV 파일 생성
     async generateCsv(allRepoScores, outputDir = '.') {
-        const promises = Array.from(allRepoScores).map(async ([repoName, repoScores]) => {
-            // 통합 CSV 생성
-            const repoActivities = this.participants.get(repoName);
-            if (repoActivities) {
-                await this.generateCombinedCsv(repoName, repoScores, repoActivities, outputDir);
-            }
-        });
-
-        await Promise.all(promises);
+        try {
+            const csvGenerator = new CsvGenerator();
+            await csvGenerator.generateCsv(allRepoScores, this.participants, outputDir);
+        } catch (error) {
+            log(`CSV 생성 중 오류 발생: ${error.message}`, 'ERROR');
+            throw error;
+        }
     }
 
     setTheme(themeName) {
